[CIT-450] Remove dependency on mqtt_client from tedge_sm
