+++ sm/tedge_sm/Cargo.toml
@@ -5,10 +5,8 @@ authors = ["diw"]
 edition = "2018"
 
 [dependencies]
-mqtt_client = { path = "../../common/mqtt_client" }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
-tedge_users ={ path = "../../common/tedge_users" }
 thiserror = "1.0"
 url = "2.2"
